requirements.txt
streamlit
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pandas
pypdf
Pillow
google-generativeai
openpyxl
